URL updates
for some reason the technical analysis api was creating a conflict with the self.URL of the classes. This one was not being replaced at the moment of calling the other apis, therefore, the api send an error message.
Now each sdk of each api have a unique URL host.